# Decoder: disallow modification of existing table